requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
chardet==4.0.0
discord-py-slash-command==1.1.0
discord.py==1.6.0
html2text==2020.1.16
idna==3.1
multidict==5.1.0
PyYAML==5.4.1
soupsieve==2.2.1
typing-extensions==3.7.4.3
wavelink==0.9.9
yarl==1.6.3
